+++ install.py
@@ -7,7 +7,7 @@
 
 from weecfg.extension import ExtensionInstaller
 
-VERSION = '0.6.0'
+VERSION = '0.7.0'
 
 
 def loader():